refactor: deduplicate cleanup run-lease primitives into cleanup-run-lease
Consolidate CLEANUP_RUN_LEASE_MS, heartbeat constants, acquire/release/renew, startCleanupRunLease, conflict error classes, and with*MutationLease wrappers into the canonical cleanup-run-lease module. Repoint executor, scheduler, routes, and tests to the canonical module.
