+++ apps/api/src/lib/library-cleanup/__tests__/shared-plex-safety.test.ts
@@ -5,10 +5,12 @@ import { PlexMovieNotFoundError } from "../../plex/plex-client.js";
 import { withQuiObservationTopologyGuard } from "../../qui/observation-topology-guard.js";
 import { appendCleanupAuditEvent, appendCleanupTerminalAuditEvent } from "../cleanup-audit.js";
 import {
-	buildCleanupPreviewDetails,
-	cleanupApprovalTargetKey,
 	CleanupRunAlreadyInProgressError,
 	CleanupRunLeaseLostError,
+} from "../cleanup-run-lease.js";
+import {
+	buildCleanupPreviewDetails,
+	cleanupApprovalTargetKey,
 	executeApprovedItems,
 	executeCleanupPreview,
 	executeCleanupRun,
