+++ apps/api/src/lib/library-cleanup/cleanup-scheduler.ts
@@ -27,14 +27,13 @@ import {
 	type CleanupTerminalAuditStatus,
 } from "./cleanup-audit.js";
 import {
-	CLEANUP_RUN_LEASE_MS,
-	CleanupRunAlreadyInProgressError,
 	executeCleanupRun,
 	INTERRUPTED_CLEANUP_RECOVERY_MESSAGE,
 	SONARR_EPISODE_UNMONITOR_CONFIRMED_RECOVERY_MESSAGE,
 	SONARR_EPISODE_UNMONITOR_PARTIAL_MESSAGE,
 	SONARR_EPISODE_UNMONITOR_STARTED_RECOVERY_MESSAGE,
 } from "./cleanup-executor.js";
+import { CLEANUP_RUN_LEASE_MS, CleanupRunAlreadyInProgressError } from "./cleanup-run-lease.js";
 import {
 	CleanupMaintenanceConflictError,
 	withCleanupOperationGuard,
