+++ apps/api/src/routes/trash-guides/__tests__/manual-arr-writer-guard.test.ts
@@ -11,7 +11,7 @@ vi.mock("../../../lib/trash-guides/deployment-operation-gate.js", () => ({
 	assertNoPendingDeploymentOperation,
 }));
 
-vi.mock("../../../lib/library-cleanup/cleanup-executor.js", () => ({
+vi.mock("../../../lib/library-cleanup/cleanup-run-lease.js", () => ({
 	withCleanupTopologyMutationLease,
 }));
